+++ benchmark/agbenchmark/utils/data_types.py
@@ -1,7 +1,4 @@
-import datetime
 import json
-import sys
-from datetime import datetime
 from enum import Enum
 from pathlib import Path
 from typing import Any, Dict, List, Optional
@@ -31,80 +28,6 @@ DIFFICULTY_MAP = {
 }
 
 STRING_DIFFICULTY_MAP = {e.value: DIFFICULTY_MAP[e] for e in DifficultyLevel}
-
-
-def calculate_info_test_path(base_path: Path, benchmark_start_time: datetime) -> Path:
-    """
-    Calculates the path to the directory where the test report will be saved.
-    """
-    # Ensure the reports path exists
-    base_path.mkdir(parents=True, exist_ok=True)
-
-    # Get current UTC date-time stamp
-    date_stamp = benchmark_start_time.strftime("%Y%m%dT%H%M%S")
-
-    # Default run name
-    run_name = "full_run"
-
-    # Map command-line arguments to their respective labels
-    arg_labels = {
-        "--test": None,
-        "--category": None,
-        "--maintain": "maintain",
-        "--improve": "improve",
-        "--explore": "explore",
-    }
-
-    # Identify the relevant command-line argument
-    for arg, label in arg_labels.items():
-        if arg in sys.argv:
-            test_arg = sys.argv[sys.argv.index(arg) + 1] if label is None else None
-            run_name = arg.strip("--")
-            if test_arg:
-                run_name = f"{run_name}_{test_arg}"
-            break
-
-    # Create the full new directory path with ISO standard UTC date-time stamp
-    report_path = base_path / f"{date_stamp}_{run_name}"
-
-    # Ensure the new directory is created
-    report_path.mkdir(exist_ok=True)
-    return report_path
-
-
-class AgentBenchmarkConfig(BaseModel):
-    """
-    This class represents the configuration for the Agent agbenchmark.
-    It includes the following attributes:
-    - agent_benchmark_config_path: The path to the agent benchmark config that this object was created from.
-    - reports_folder: The path to the folder where the benchmark reports will be stored.
-    - host: The host where the benchmark is run.
-    """
-
-    agent_benchmark_config_path: Path | None = None
-    reports_folder: Path | None = None
-    host: str | None
-
-    def get_reports_location(self) -> Path:
-        # if not self.reports_folder:
-        #     self.reports_folder = (
-        #         Path(self.agent_benchmark_config_path).parent / "reports"
-        #     ).resolve()
-        return Path.cwd() / "agbenchmark_config" / "reports"
-
-    def get_reports_path(self, benchmark_start_time: datetime) -> Path:
-        return calculate_info_test_path(
-            self.get_reports_location(), benchmark_start_time
-        )
-
-    def get_regression_reports_path(self) -> Path:
-        return self.get_reports_location() / "regression_tests.json"
-
-    def get_success_rate_path(self) -> Path:
-        return self.get_reports_location() / "success_rate.json"
-
-    def get_agent_home_directory(self) -> Path:
-        return Path(self.agent_benchmark_config_path).resolve().parent
 
 
 class Info(BaseModel):
